# api-docs.json
{"openapi":"3.0.1","info":{"title":"Equipment Service API - Gimnasio","description":"API para gestionar equipos en el gimnasio","version":"1.0.0"},"servers":[{"url":"http://localhost:8083","description":"Generated server url"}],"tags":[{"name":"Gestión de Equipos","description":"Operaciones para administrar equipos del gimnasio"}],"paths":{"/{id}":{"get":{"tags":["Gestión de Equipos"],"summary":"Obtener un equipo por ID","description":"Devuelve un equipo por su ID. Accesible para ADMIN, TRAINER y MEMBER.","operationId":"getEquipmentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Acceso denegado por falta de permisos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"200":{"description":"Equipo obtenido con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"404":{"description":"Equipo no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}}}},"put":{"tags":["Gestión de Equipos"],"summary":"Actualizar un equipo","description":"Actualiza un equipo existente. Solo para ROLE_ADMIN.","operationId":"updateEquipment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipment"}}},"required":true},"responses":{"403":{"description":"Permisos insuficientes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"200":{"description":"Equipo actualizado con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"404":{"description":"Equipo no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}}}},"delete":{"tags":["Gestión de Equipos"],"summary":"Eliminar un equipo","description":"Elimina un equipo existente. Solo para ROLE_ADMIN.","operationId":"deleteEquipment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Permisos insuficientes"},"204":{"description":"Equipo eliminado con éxito"},"404":{"description":"Equipo no encontrado"}}}},"/":{"get":{"tags":["Gestión de Equipos"],"summary":"Listar todos los equipos","description":"Devuelve una lista de equipos. Accesible para ADMIN, TRAINER y MEMBER.","operationId":"getAllEquipment_1","responses":{"200":{"description":"Lista de equipos obtenida con éxito","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Equipment"}}}}},"403":{"description":"Acceso denegado por falta de permisos","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Equipment"}}}}}}},"post":{"tags":["Gestión de Equipos"],"summary":"Crear un nuevo equipo","description":"Registra un equipo nuevo. Solo para ROLE_ADMIN.","operationId":"addEquipment_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipment"}}},"required":true},"responses":{"403":{"description":"Permisos insuficientes","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"200":{"description":"Equipo creado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}},"400":{"description":"Datos inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Equipment"}}}}}}}},"components":{"schemas":{"Equipment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"condition":{"type":"string"}}}}}}